feat: full wire-up
