+++ src/lib/swapper/swappers/ThorchainSwapper/getThorTradeQuote/getTradeQuote.ts
@@ -9,10 +9,7 @@ import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingl
 import { baseUnitToPrecision, bn, bnOrZero, convertPrecision } from 'lib/bignumber/bignumber'
 import { fromBaseUnit, toBaseUnit } from 'lib/math'
 import { getThorTxInfo as getEvmThorTxInfo } from 'lib/swapper/swappers/ThorchainSwapper/evm/utils/getThorTxData'
-import {
-  DEFAULT_STREAMING_INTERVAL,
-  THORCHAIN_FIXED_PRECISION,
-} from 'lib/swapper/swappers/ThorchainSwapper/utils/constants'
+import { THORCHAIN_FIXED_PRECISION } from 'lib/swapper/swappers/ThorchainSwapper/utils/constants'
 import { getQuote } from 'lib/swapper/swappers/ThorchainSwapper/utils/getQuote/getQuote'
 import { getUtxoTxFees } from 'lib/swapper/swappers/ThorchainSwapper/utils/txFeeHelpers/utxoTxFees/getUtxoTxFees'
 import { getThorTxInfo as getUtxoThorTxInfo } from 'lib/swapper/swappers/ThorchainSwapper/utxo/utils/getThorTxData'
@@ -36,8 +33,10 @@ import {
 } from 'state/slices/tradeQuoteSlice/utils'
 
 import { THORCHAIN_STREAM_SWAP_SOURCE } from '../constants'
-import type { ThornodeQuoteResponseSuccess } from '../types'
+import type { ThornodePoolResponse, ThornodeQuoteResponseSuccess } from '../types'
 import { addSlippageToMemo } from '../utils/addSlippageToMemo'
+import { assetIdToPoolAssetId } from '../utils/poolAssetHelpers/poolAssetHelpers'
+import { thorService } from '../utils/thorService'
 import { getEvmTxFees } from '../utils/txFeeHelpers/evmTxFees/getEvmTxFees'
 
 export type ThorEvmTradeQuote = TradeQuote &
@@ -106,6 +105,49 @@ export const getThorTradeQuote = async (
     affiliateBps: requestedAffiliateBps,
   })
 
+  const daemonUrl = getConfig().REACT_APP_THORCHAIN_NODE_URL
+  const maybePoolsResponse = await thorService.get<ThornodePoolResponse[]>(
+    `${daemonUrl}/lcd/thorchain/pools`,
+  )
+
+  if (maybePoolsResponse.isErr()) return Err(maybePoolsResponse.unwrapErr())
+
+  const { data: poolsResponse } = maybePoolsResponse.unwrap()
+
+  const buyPoolId = assetIdToPoolAssetId({ assetId: buyAsset.assetId })
+  const sellPoolId = assetIdToPoolAssetId({ assetId: sellAsset.assetId })
+  const sellAssetPool = poolsResponse.find(pool => pool.asset === sellPoolId)
+  const buyAssetPool = poolsResponse.find(pool => pool.asset === buyPoolId)
+
+  if (!sellAssetPool)
+    return Err(
+      makeSwapErrorRight({
+        message: `[_getQuote]: Pool not found for sell asset ${sellAsset.assetId}`,
+        code: SwapErrorType.POOL_NOT_FOUND,
+        details: { sellAssetId: sellAsset.assetId, buyAssetId: buyAsset.assetId },
+      }),
+    )
+
+  if (!buyAssetPool)
+    return Err(
+      makeSwapErrorRight({
+        message: `[_getQuote]: Pool not found for buy asset ${buyAsset.assetId}`,
+        code: SwapErrorType.POOL_NOT_FOUND,
+        details: { sellAssetId: sellAsset.assetId, buyAsset: buyAsset.assetId },
+      }),
+    )
+
+  const sellAssetDepthBps = sellAssetPool.derived_depth_bps
+  const buyAssetDepthBps = buyAssetPool.derived_depth_bps
+  const swapDepthBps = bn(sellAssetDepthBps).plus(buyAssetDepthBps).div(2)
+
+  const streamingInterval = (() => {
+    // Low health for the pools of this swap - use a longer streaming interval
+    if (swapDepthBps.lt(5000)) return 10
+    if (swapDepthBps.lt(9000) && swapDepthBps.gte(5000)) return 5
+    return 1
+  })()
+
   const maybeStreamingSwapQuote = thorSwapStreamingSwaps
     ? await getQuote({
         sellAsset,
@@ -114,6 +156,7 @@ export const getThorTradeQuote = async (
         receiveAddress,
         streaming: true,
         affiliateBps: requestedAffiliateBps,
+        streamingInterval,
       })
     : undefined
 
@@ -237,7 +280,7 @@ export const getThorTradeQuote = async (
               chainId: sellAsset.chainId,
               affiliateBps,
               isStreaming,
-              streamingInterval: DEFAULT_STREAMING_INTERVAL,
+              streamingInterval,
             })
             const { data, router } = await getEvmThorTxInfo({
               sellAsset,
@@ -320,7 +363,7 @@ export const getThorTradeQuote = async (
               isStreaming,
               chainId: sellAsset.chainId,
               affiliateBps,
-              streamingInterval: DEFAULT_STREAMING_INTERVAL,
+              streamingInterval,
             })
             const { vault, opReturnData, pubkey } = await getUtxoThorTxInfo({
               sellAsset,
@@ -417,7 +460,7 @@ export const getThorTradeQuote = async (
               isStreaming,
               chainId: sellAsset.chainId,
               affiliateBps,
-              streamingInterval: DEFAULT_STREAMING_INTERVAL,
+              streamingInterval,
             })
 
             return {
