+++ src/lib/swapper/swappers/ThorchainSwapper/utils/test-data/responses.ts
@@ -18,6 +18,7 @@ export const btcThornodePool: ThornodePoolResponse = {
   synth_supply: '44718623180',
   synth_supply_remaining: '8963246519',
   synth_units: '142965023220851',
+  derived_depth_bps: 8368,
 }
 
 export const ethThornodePool: ThornodePoolResponse = {
@@ -35,6 +36,7 @@ export const ethThornodePool: ThornodePoolResponse = {
   synth_supply: '218601192670',
   synth_supply_remaining: '183106737474',
   synth_units: '36567637964216',
+  derived_depth_bps: 8298,
 }
 
 export const foxThornodePool: ThornodePoolResponse = {
@@ -52,6 +54,7 @@ export const foxThornodePool: ThornodePoolResponse = {
   synth_supply: '3126453386967',
   synth_supply_remaining: '161390775916374',
   synth_units: '50300400994',
+  derived_depth_bps: 0,
 }
 
 export const usdcThornodePool: ThornodePoolResponse = {
@@ -69,6 +72,7 @@ export const usdcThornodePool: ThornodePoolResponse = {
   synth_supply: '20420358430864',
   synth_supply_remaining: '102379864074658',
   synth_units: '1423205664974',
+  derived_depth_bps: 0,
 }
 
 export const thornodePools: ThornodePoolResponse[] = [
